+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Use arrayvec for PrincipalVariation (15.52 +- 8.72)
 * Fix an accidental fail hard in quiescence - fail soft when eval >= beta (12.54 +- 7.15)
+* Add futility pruning (10.82 +- 6.41)
 * Pack Move into u16 and store extra information, e.g. if the move is a capture (7.03 +- 4.89)
 * Add tablebase support and follow tablebase lines (6.79 +- 4.59 Elo (5-man vs none))
 * Use arrayvec for MoveList (1.05 +- 3.35)